+++ src/containers/EditListingPage/EditListingWizard/EditListingDetailsPanel/EditListingDetailsForm.test.js
@@ -4,9 +4,9 @@ import { fakeIntl } from '../../../../util/test-data';
 import EditListingDetailsForm from './EditListingDetailsForm';
 
 const noop = () => null;
-const selectableTransactionTypes = [
+const selectableListingTypes = [
   {
-    transactionType: 'sell-bicycles',
+    listingType: 'sell-bicycles',
     transactionProcessAlias: 'default-buying-products/release-1',
     unitType: 'item',
   },
@@ -16,7 +16,7 @@ const listingExtendedDataConfig = [
   {
     key: 'category',
     scope: 'public',
-    includeForTransactionTypes: ['sell-bicycles'],
+    includeForListingTypes: ['sell-bicycles'],
     schemaType: 'enum',
     schemaOptions: [
       { option: 'men', label: 'Men' },
@@ -38,7 +38,7 @@ const listingExtendedDataConfig = [
   {
     key: 'amenities',
     scope: 'public',
-    includeForTransactionTypes: [
+    includeForListingTypes: [
       'rent-bicycles-daily',
       'rent-bicycles-nightly',
       'rent-bicycles-hourly',
@@ -76,8 +76,8 @@ describe('EditListingDetailsForm', () => {
         disabled={false}
         ready={false}
         listingExtendedDataConfig={listingExtendedDataConfig}
-        selectableTransactionTypes={selectableTransactionTypes}
-        hasExistingTransactionType={true}
+        selectableListingTypes={selectableListingTypes}
+        hasExistingListingType={true}
       />
     );
     expect(tree).toMatchSnapshot();