+++ src/containers/EditListingPage/EditListingWizard/EditListingWizardTab.js
@@ -169,7 +169,7 @@ const EditListingWizardTab = props => {
       ready: newListingPublished,
       disabled: fetchInProgress,
       submitButtonText: tabSubmitButtonText,
-      transactionTypes: config.transaction.transactionTypes,
+      listingTypes: config.listing.listingTypes,
       onManageDisableScrolling,
       onSubmit: values => {
         return onCompleteEditListingWizardTab(tab, values);